// frontend/admin/src/shared/types/generated.d.ts
export interface components {
    schemas: {
// ...
        AdminUpdateRequest: {
            username: string;
            role: string;
            enabled: boolean;
        };
// ...
        AdminCreateRequest: {
            username: string;
            password: string;
            role: string;
            enabled: boolean;
        };
// ...
        AdminResponse: {
            /** Format: uuid */
            id?: string;
            username?: string;
            role?: string;
            enabled?: boolean;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            updatedAt?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}